CloudFunctions/requirements.txt
functions-framework==3.*
wbgapi
pandas
google-cloud-bigquery[pandas]
google-cloud-storage
pyarrow
pyjstat
numpy
statsmodels